Create sample.xml
A sample that was at a "5 finger discount". (Stolen from the Blogger. I tried to ask permission but the author's' website no longer exists.)
This XML file is provided by Blogger, so it's meant to be distributed. That's all I know.

This is being stripped down to it's bare minimum, then the Zurb Foundation 6 structure will be adhered to for the new example.